requirements.txt
Flask
pymupdf
Pillow
reportlab
